ref: 紙吹雪の処理をBoardコンポーネントから分離する
