+++ src/app/rooms/[id]/_components/Board.tsx
@@ -1,12 +1,10 @@
 'use client';
 
-import { useState, useEffect } from 'react';
-import { ref, update, onValue, remove } from 'firebase/database';
+import { ref, update, remove } from 'firebase/database';
 import { RefreshCw, Eye } from 'lucide-react';
-import Confetti from 'react-confetti';
 import { db } from '@/lib/firebase';
-import { Room } from '@/types';
 import { useAuth } from '@/hooks/useAuth';
+import { useRoomData } from '@/hooks/useRoomData';
 import { PlayerCard } from './PlayerCard';
 
 type Props = {
@@ -14,33 +12,8 @@ type Props = {
 };
 
 export function Board({ roomId }: Props) {
-  const [roomData, setRoomData] = useState<Room | null>(null);
-  const { user, loading: authLoading } = useAuth();
-  const [windowSize, setWindowSize] = useState({ width: 0, height: 0 });
-
-  // ウィンドウサイズの取得
-  useEffect(() => {
-    const updateWindowSize = () => {
-      setWindowSize({ width: window.innerWidth, height: window.innerHeight });
-    };
-    updateWindowSize();
-    window.addEventListener('resize', updateWindowSize);
-    return () => window.removeEventListener('resize', updateWindowSize);
-  }, []);
-
-  // リアルタイム同期
-  useEffect(() => {
-    // 認証前の場合はデータ取得ができないためリアルタイム同期を停止
-    if (!user) return;
-
-    const roomRef = ref(db, `rooms/${roomId}`);
-    const unsubscribe = onValue(roomRef, (snapshot) => {
-      const data = snapshot.val() as Room | null;
-      setRoomData(data);
-    });
-
-    return () => unsubscribe();
-  }, [roomId, user]);
+  const { user } = useAuth();
+  const { roomData, isLoading } = useRoomData(roomId);
 
   const usersList = Object.entries(roomData?.users || {});
   const revealed = roomData?.status === 'revealed';
@@ -58,14 +31,6 @@ export function Board({ roomId }: Props) {
       ? (validVotes.reduce((sum, vote) => sum + vote, 0) / validVotes.length).toFixed(1)
       : null;
 
-  // 全員の投票が一致しているかチェック
-  const allVotesMatch =
-    revealed &&
-    validVotes.length >= 2 &&
-    validVotes.every((vote) => vote === validVotes[0]);
-
-  const isLoading = authLoading || !roomData;
-
   const handleRemovePlayer = (targetUid: string) => {
     if (!user || !roomId || targetUid === user.uid) return;
     const userRef = ref(db, `rooms/${roomId}/users/${targetUid}`);
@@ -73,17 +38,7 @@ export function Board({ roomId }: Props) {
   };
 
   return (
-    <>
-      {allVotesMatch && windowSize.width > 0 && windowSize.height > 0 && (
-        <Confetti
-          width={windowSize.width}
-          height={windowSize.height}
-          recycle={false}
-          numberOfPieces={500}
-          gravity={0.3}
-        />
-      )}
-      <div className="bg-white/80 backdrop-blur-sm rounded-2xl shadow-xl p-4 sm:p-6 mb-6 sm:mb-8">
+    <div className="bg-white/80 backdrop-blur-sm rounded-2xl shadow-xl p-4 sm:p-6 mb-6 sm:mb-8">
       {/* 上部ステータスエリア */}
       <div className="mb-4 sm:mb-6 space-y-2 sm:space-y-3 text-center">
         {isLoading ? (
@@ -163,7 +118,6 @@ export function Board({ roomId }: Props) {
         ))}
       </div>
     </div>
-    </>
   );
 }
